Matherial Design, Dark and light mode, Dynamic layout expanding description, set favourite words, progress bar, no data design, ic_launcher

+++ app/src/main/java/uz/gita/justdictionary/domain/repository/AppRepository.kt
@@ -6,8 +6,8 @@ import kotlinx.coroutines.flow.Flow
 interface AppRepository {
     suspend fun getAllWords(): Cursor
 
-    suspend fun rememberWord(id: Int)
-    suspend fun forgetWord(id:Int)
+    suspend fun rememberWord(id: Long)
+    suspend fun forgetWord(id: Long)
 
     suspend fun searchWord(query: String): Cursor
 }
