+++ app/src/main/java/uz/gita/justdictionary/domain/repository/impl/AppRepositoryImpl.kt
@@ -1,8 +1,6 @@
 package uz.gita.justdictionary.domain.repository.impl
 
 import android.database.Cursor
-import android.util.Log
-import uz.gita.justdictionary.data.local.AppDatabase
 import uz.gita.justdictionary.data.local.dao.DictionaryDao
 import uz.gita.justdictionary.domain.repository.AppRepository
 import javax.inject.Inject
@@ -13,11 +11,11 @@ class AppRepositoryImpl @Inject constructor(
     override suspend fun getAllWords(): Cursor = dao.getAllWords()
     override suspend fun searchWord(query: String): Cursor = dao.searchWord(query)
 
-    override suspend fun rememberWord(id: Int) {
+    override suspend fun rememberWord(id: Long) {
         dao.rememberWord(id)
     }
 
-    override suspend fun forgetWord(id: Int) {
+    override suspend fun forgetWord(id: Long) {
         dao.forgetWord(id)
     }
